Modify setting and getting of card names.
Set and get newly created cards names by using jQuerys .data() function.
This is in order to set up for retrieving this card specific data (and later other data, like deadline and description) for dynamic output to the card dialog.
Otherwise the data set in the card dialog is not saved to a specific element, and can be overwritten when trying to change another cards data.